refactor(mcp): colocate connector catalog UI
