+++ airbyte/mcp/interactive/__init__.py
@@ -5,9 +5,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from fastmcp_extensions import register_mcp_tools
-
-from airbyte.mcp._ui_builders import register_prefab_tool_metadata
+from airbyte.mcp._tool_utils import register_mcp_prefab_tools
 from airbyte.mcp.interactive._registry import show_connectors_list
 
 
@@ -20,8 +18,7 @@ INTERACTIVE_TOOL_NAMES = frozenset({"show_connectors_list"})
 
 def register_interactive_tools(app: FastMCP) -> None:
     """Register UI-presenting tools."""
-    register_mcp_tools(app, mcp_module="_registry")
-    register_prefab_tool_metadata(app, tuple(INTERACTIVE_TOOL_NAMES))
+    register_mcp_prefab_tools(app, mcp_module="_registry")
 
 
 __all__ = [
